api-gateway/requirements.txt
Flask
Werkzeug==2.2.2
scipy>1.6.2
pandas>1.2.4
flask_pymongo
msgpack==1.0.2
flask-SQLAlchemy
psycopg2-binary>=2.8
datetime
streamlit